# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Products API",
        "description": "API para gestión de productos con soporte multi-moneda",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/products/{product_id}/prices": {
            "get": {
                "tags": [
                    "Precios de Productos"
                ],
                "summary": "Listar precios del producto",
                "description": "Retorna todos los precios de un producto en diferentes monedas",
                "operationId": "76b88800098f0b2d197bc306ecb356d6",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Precios del producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "base_price": {
                                                    "type": "object"
                                                },
                                                "prices": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Precios de Productos"
                ],
                "summary": "Crear/actualizar precio",
                "description": "Crea o actualiza el precio de un producto en una moneda específica",
                "operationId": "bed13ea51028e5d85b0ddd051cf8aa18",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "currency_id",
                                    "price"
                                ],
                                "properties": {
                                    "currency_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": 89.99
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Precio creado"
                    },
                    "200": {
                        "description": "Precio actualizado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/{product_id}/prices/{price_id}": {
            "delete": {
                "tags": [
                    "Precios de Productos"
                ],
                "summary": "Eliminar precio",
                "description": "Elimina un precio específico de un producto",
                "operationId": "d4601e371cfc6fdd11d90362bce27837",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "price_id",
                        "in": "path",
                        "description": "ID del precio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Precio eliminado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Recurso no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingresa tu token Bearer",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}